homework7

+++ Tasks/HomeWork/Task7/src/com/company/Rectangle.java
@@ -2,42 +2,42 @@ package com.company;
 
 public class Rectangle extends Square {
 
-    int b;
+    double b;
 
     Rectangle() {
 
         b = 7;
     }
 
-    Rectangle(int a, int b) {
+    Rectangle(double a, double b) {
         this.a = a;
         this.b = b;
     }
 
-    public void setB(int b) {
+    public void setB(double b) {
         this.b = b;
     }
 
-    public void setA(int a) {
+    public void setA(double a) {
         this.a = a;
     }
 
-    public int getB() {
+    public double getB() {
         return b;
     }
 
-    public int getA() {
+    public double getA() {
         return a;
     }
 
-    public void Squareds(int a, int b) {
+    public void Squareds(double a, double b) {
         if (a > 0 && b > 0) {
             System.out.println("Площадь прямоугольника = " + (a + b) * 2);
         } else System.err.println("Площадь прямоугольника не может быть нулевым");
 
     }
 
-    public void Perimetr(int a, int b) {
+    public void Perimetr(double a, double b) {
         if (a > 0 && b > 0) {
             System.out.println("Площадь прямоугольника = " + (a + b) * 2);
         } else System.err.println("Площадь прямоугольника не может быть нулевым");
